Hoist common unknown/deleted label handling
This allows us to set the right return type for the underlying
database util and makes handling of unknown labels in process_gate
and process_action explicit (though admitedly in those cases it
shouldn't ever happen).
