+++ routemaster/state_machine/gates.py
@@ -9,7 +9,6 @@ from routemaster.state_machine.utils import (
     get_label_metadata,
     get_current_history,
 )
-from routemaster.state_machine.exceptions import DeletedLabel
 
 
 def process_gate(
@@ -35,9 +34,7 @@ def process_gate(
 
     gate = state
 
-    metadata, deleted = get_label_metadata(app, label, state_machine)
-    if deleted:
-        raise DeletedLabel(label)
+    metadata = get_label_metadata(app, label)
 
     history_entry = get_current_history(app, label)
